Add CS50 from Harvard
Simply the best. Learning by doing is the way. Some tasks can be hard for absolute beginners,
but with a little patience, thinking and following great instructors makes overcoming hard tasks a
great pleasure. And it's free (without the certificate).
Projects range from classic 'Hello World' in C to Python web app using Flask and SQLite back-end and JavaScript.
A lot of the code for bigger projects is already written by staff, so it gives some real-world,
reading and working with other people's code experience.